Sun, Nov 13, 2022, 10:25 PM +00:00

+++ finance/app.py
@@ -50,7 +50,7 @@ def index():
     print(stocks)
     for stock in stocks:
         stock.update({"pricenow": lookup(stock['symbol'])['price']})
-        c = c + stock['pricenow']*stock['SUM(number)']
+        c = c + float(stock['pricenow']*stock['SUM(number)'])
     total = c
     bank = db.execute("SELECT cash FROM users WHERE id = ?", session["user_id"])[0].get("cash")
     return render_template("index.html", stocks=stocks, bank=bank, total = total)
